## Why is this change being made?
Adding SCP Heartbeat test on FPGA

## What is changed?
- Added infrastructure to run pythia tests on FPGA:
   1. Updated folder structures to support both SVP and FPGA tests. The steps and modules needed for FPGA and SVP are different atm, so split them into 2 folders for both robot files as well as python scripts.
   1. Added hosts.json file for FPGA SCP heartbeat as per Pythia 2.0 requirements.
   1. Ingested Kingsgate.FW.Scripts to load prodFW on FPGA. Added wrapper cmm script to direct scripts to appropriate elfs.
   1. Minor updates and cleanup in repo structure
   1. Updated pipeline (*.yml) files as needed.

## How is the change being tested?

- [ ] No test, document update only
- [ ] Unit tests pass
- [ ] Integration tests pass
- [x] Manual tests pass (add test specifics): Ran FPGA pipeline manually and ensure heartbeat test passes
- [x] Others, explain: Ensure all CI tests pass for SVP for CI pipeline without any impact

### Test Platform

- [x] SVP
- [x] FPGA (R19)
- [ ] Silicon
- [ ] Others, explain:

## Notes
Pipeline Link: https://azurecsi.visualstudio.com/Woodinville/_build/results?buildId=693800&view=results

## Risks

Related work items: #1885236

+++ tools/pwsh/modules/Pythia-Utils.psm1
@@ -43,7 +43,8 @@ Path to the test to run. Can be a robot file, or a directory containing robot fi
 Invoke-Pythia -test .\tests\functional\pythia\test_suites\heart_beat\
 #>
 Function Invoke-Pythia(
-    [Parameter(Mandatory=$true)] [string] $test
+    [Parameter(Mandatory=$true)] [string] $test,
+    [Parameter(Mandatory=$true)] [ValidateSet('svp', 'fpga')] [string] $platform = "svp"
 )
 {
 
@@ -75,7 +76,11 @@ Function Invoke-Pythia(
     # Create new json files with expanded string values, expanding environment variables
     Expand-File -in_file $workspace_config_in -out_file $workspace_config_out
 
-    $host_configs_path = Join-Path -Path $env:REPO_APP_ROOT -ChildPath "tests\functional\pythia\hosts\"
+    if ($platform -eq "svp") {
+        $host_configs_path = Join-Path -Path $env:REPO_APP_ROOT -ChildPath "tests\functional\pythia\hosts\svp_hosts"
+    } elseif ($platform -eq "fpga") {
+        $host_configs_path = Join-Path -Path $env:REPO_APP_ROOT -ChildPath "tests\functional\pythia\hosts\fpga_hosts"
+    }
 
     $iteration = 1
     Get-ChildItem $host_configs_path |
@@ -130,7 +135,9 @@ Function Invoke-Pythia(
     Write-Host "`tPythia Tests Completed. Please see tests results: $test_results_dir"
     Write-Host ""
 
-    Write-Host "Removing .svp_simulator dir"
-    Remove-Item $test_results_dir/.svp_simulator -Recurse -Force
+    if ($platform -eq "svp") {
+        Write-Host "Removing .svp_simulator dir"
+        Remove-Item $test_results_dir/.svp_simulator -Recurse -Force
+    }
 
 }
